refactor: common zod error handler
