+++ packages/backend/src/apps/gathersg/actions/create-case/index.ts
@@ -9,10 +9,11 @@ import logger from '@/helpers/logger'
 import { ensureZodEnumValue } from '@/helpers/zod-utils'
 
 import { fetchCaseFields } from '../../common/fetch-case-fields'
-import throwGatherSGStepError from '../../common/throw-errors'
+import { fieldTypeEnum } from '../../common/schema'
+import { throwGatherSGStepError } from '../../common/throw-errors'
 
 import getDataOutMetadata from './get-data-out-metadata'
-import { fieldTypeEnum, requestSchema, responseSchema } from './schema'
+import { requestSchema, responseSchema } from './schema'
 
 const action: IRawAction = {
   name: 'Create case',
