+++ packages/backend/src/apps/gathersg/actions/tag-or-untag-case/index.ts
@@ -1,12 +1,14 @@
 import type { IRawAction } from '@plumber/types'
 
 import { ZodError } from 'zod'
-import { fromZodError } from 'zod-validation-error'
 
 import HttpError from '@/errors/http'
-import StepError, { GenericSolution } from '@/errors/step'
+import StepError from '@/errors/step'
 
-import throwGatherSGStepError from '../../common/throw-errors'
+import {
+  handleZodError,
+  throwGatherSGStepError,
+} from '../../common/throw-errors'
 
 import { requestSchema, responseSchema } from './schema'
 
@@ -68,11 +70,7 @@ const action: IRawAction = {
       })
     } catch (error) {
       if (error instanceof ZodError) {
-        const firstError = fromZodError(error).details[0]
-        throw new StepError(
-          `${firstError.message}`,
-          GenericSolution.ReconfigureInvalidField,
-        )
+        handleZodError(error)
       }
 
       if (error instanceof HttpError) {
