+++ packages/backend/src/apps/gathersg/common/schema.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod'
+
+export const fieldTypeEnum = z.enum(['string', 'number', 'null'])
